Adds better formatting for the decoded certificate.
Removes the openTextDocument to view the decoded content.

Adds a virtual document to view the contents as a preview. This prevents users being prompted to save the untitled file before closing.

Adds information to the package.json file for publishing.

Adds the capability to view files not in the workspace with the peek509 decodeCert command used from the command pallet.

Adds the correct types for the DecodedCert.publicKey field.

Adds an icon for the extension.